# Speed up prefix sums when decoding doc IDs.