--- a/lucene/core/src/java/org/apache/lucene/codecs/lucene912/PForUtil.java
+++ b/lucene/core/src/java/org/apache/lucene/codecs/lucene912/PForUtil.java
@@ -130,7 +130,7 @@ final class PForUtil {
       in.readVLong();
       in.skipBytes((numExceptions << 1));
     } else {
-      in.skipBytes(forUtil.numBytes(bitsPerValue) + (numExceptions << 1));
+      in.skipBytes(ForUtil.numBytes(bitsPerValue) + (numExceptions << 1));
     }
   }
 }