--- a/lucene/core/src/java/org/apache/lucene/codecs/lucene912/PForUtil.java
+++ b/lucene/core/src/java/org/apache/lucene/codecs/lucene912/PForUtil.java
@@ -21,7 +21,6 @@ import java.util.Arrays;
 import org.apache.lucene.internal.vectorization.PostingDecodingUtil;
 import org.apache.lucene.store.DataInput;
 import org.apache.lucene.store.DataOutput;
-import org.apache.lucene.store.IndexInput;
 import org.apache.lucene.util.LongHeap;
 import org.apache.lucene.util.packed.PackedInts;
 
@@ -106,18 +105,18 @@ final class PForUtil {
   }
 
   /** Decode 128 integers into {@code ints}. */
-  void decode(IndexInput in, PostingDecodingUtil inUtil, long[] longs) throws IOException {
-    final int token = Byte.toUnsignedInt(in.readByte());
+  void decode(PostingDecodingUtil pdu, long[] longs) throws IOException {
+    final int token = Byte.toUnsignedInt(pdu.in.readByte());
     final int bitsPerValue = token & 0x1f;
     final int numExceptions = token >>> 5;
     if (bitsPerValue == 0) {
-      Arrays.fill(longs, 0, ForUtil.BLOCK_SIZE, in.readVLong());
+      Arrays.fill(longs, 0, ForUtil.BLOCK_SIZE, pdu.in.readVLong());
     } else {
-      forUtil.decode(bitsPerValue, in, inUtil, longs);
+      forUtil.decode(bitsPerValue, pdu, longs);
     }
     for (int i = 0; i < numExceptions; ++i) {
-      longs[Byte.toUnsignedInt(in.readByte())] |=
-          Byte.toUnsignedLong(in.readByte()) << bitsPerValue;
+      longs[Byte.toUnsignedInt(pdu.in.readByte())] |=
+          Byte.toUnsignedLong(pdu.in.readByte()) << bitsPerValue;
     }
   }