… lucene/CHANGES.txt
… lucene/benchmark-jmh/src/java/org/apache/lucene/benchmark/jmh/PostingIndexInputBenchmark.java
… lucene/core/src/generated/checksums/generateForDeltaUtil.json
… lucene/core/src/generated/checksums/generateForUtil.json
… lucene/core/src/java/org/apache/lucene/codecs/lucene912/gen_ForDeltaUtil.py
… lucene/core/src/java/org/apache/lucene/codecs/lucene912/gen_ForUtil.py
… lucene/core/src/java/org/apache/lucene/internal/vectorization/DefaultVectorizationProvider.java
… lucene/core/src/java/org/apache/lucene/internal/vectorization/PostingDecodingUtil.java
… lucene/core/src/java/org/apache/lucene/internal/vectorization/VectorizationProvider.java
… lucene/core/src/java21/org/apache/lucene/internal/vectorization/MemorySegmentPostingDecodingUtil.java
… lucene/core/src/java21/org/apache/lucene/internal/vectorization/PanamaVectorizationProvider.java
… lucene/core/src/test/org/apache/lucene/codecs/lucene912/TestForDeltaUtil.java
… lucene/core/src/test/org/apache/lucene/codecs/lucene912/TestForUtil.java
… lucene/core/src/test/org/apache/lucene/codecs/lucene912/TestPForUtil.java
… lucene/core/src/test/org/apache/lucene/internal/vectorization/TestPostingDecodingUtil.java