@jpountz 2024-08-17
Sadly, this seems to have caused a minor regression on nightly benchmarks, e.g. [`AndHighMed`](https://people.apache.org/~mikemccand/lucenebench/AndHighMed.html) lost 2% with a p-value of 0.

I cannot reproduce the slowdown locally. However, one thing I am observing is that `splitLongs` is never inlined into the `decodeX` methods, which is annoying as I would expect it to be important performance-wise to unroll various loops. I am not sure if this is due to insufficient warming or if we could organize the code a bit differently to help the JVM make better inlining decisions.