@jpountz 2024-08-19
Thanks for looking @gsmiller ! Are the splitLongs calls also not inlined for you as well? I submitted a PR to make nightly reports tell us about which methods get inlined / compiled / interpreted: https://github.com/apache/lucene/pull/13670 as this sounds like useful information for performance analysis.

I'm currently playing with small refactorings to see if I can get this code inlined without bumping `-XX:InlineSmallCode=2500` but I'm not successful so far.