@gsmiller 2024-08-19
Good question @jpountz. If I'm looking at this properly, it seems to not be getting inlined in my benchmark run. (I grabbed your patch from #13670 and loaded one of the jfr files produced from luceneutil):

```
PERCENT       CPU SAMPLES   STACK
[...]
0.05%         5             org.apache.lucene.internal.vectorization.MemorySegmentPostingDecodingUtil#splitLongs() [JIT compiled code]
```