@gsmiller 2024-08-26
@jpountz I'm in favor of keeping the change as well (your reasoning here makes sense to me). I'll see if I can get another data point by running our internal Amazon product search benchmarks as well to see if anything interesting shows up there on a different workload (not implying that should be a deciding factor by any means, just trying to think of ways to get additional information). If I'm able to find time for that this week I'll report back.